+++ extract_msg/attachments/attachment.py
@@ -216,6 +216,6 @@ class Attachment(AttachmentBase):
     @property
     def type(self) -> AttachmentType:
         """
-        Returns the (internally used) type of the data.
+        Returns an enum value that identifies the type of attachment.
         """
         return AttachmentType.DATA