+++ extract_msg/attachments/broken_att.py
@@ -29,6 +29,6 @@ class BrokenAttachment(AttachmentBase):
     @property
     def type(self) -> AttachmentType:
         """
-        Returns the (internally used) type of the data.
+        Returns an enum value that identifies the type of attachment.
         """
         return AttachmentType.BROKEN