+++ extract_msg/attachments/custom_att.py
@@ -195,6 +195,6 @@ class CustomAttachment(AttachmentBase):
     @property
     def type(self) -> AttachmentType:
         """
-        Returns the (internally used) type of the data.
+        Returns an enum value that identifies the type of attachment.
         """
         return AttachmentType.CUSTOM