+++ extract_msg/attachments/emb_msg_att.py
@@ -129,6 +129,6 @@ class EmbeddedMsgAttachment(AttachmentBase):
     @property
     def type(self) -> AttachmentType:
         """
-        Returns the (internally used) type of the data.
+        Returns an enum value that identifies the type of attachment.
         """
         return AttachmentType.MSG