+++ extract_msg/attachments/unsupported_att.py
@@ -35,6 +35,6 @@ class UnsupportedAttachment(AttachmentBase):
     @property
     def type(self) -> AttachmentType:
         """
-        Returns the (internally used) type of the data.
+        Returns an enum value that identifies the type of attachment.
         """
         return AttachmentType.UNSUPPORTED